Extract out code into more specific classses
