+++ duke.txt
@@ -1,4 +1,2 @@
 [T][ ] borrow book
 [D][ ] return book /by 02/12/2019 1800
-[D][ ] return book /by 02/12/2019 1800
-[E][ ] project meeting /from 04/12/2019 1400 /to 04/12/2019 1400
