+++ src/main/java/Driver.java
@@ -1,0 +1,8 @@
+import java.util.ArrayList;
+
+public class Driver {
+
+    public static void main (String [] args) {
+        Dukey dukey = new Dukey();
+    }
+}
